% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCytometrySPoTApi.r
\name{FlowCytometrySPoTApi_search_groups}
\alias{FlowCytometrySPoTApi_search_groups}
\title{Retrieve groups that match a query}
\usage{
FlowCytometrySPoTApi_search_groups(
  filter,
  query,
  search_specific_terms,
  use_versions,
  returned_metadata_fields,
  page_offset,
  page_limit,
  ...
)
}
\arguments{
\item{filter}{Filter by flow cytometry metadata (key-value metadata pair(s)). E.g. \code{Organ=blood}}

\item{query}{Search for flow cytometry objects via a full text query over all flow cytometry metadata.}

\item{search_specific_terms}{If the full-text query term is present in an ODM dictionary, enabling this parameter will modify the query to include child terms of the full-text query.  For example, the search query "Body fluid" can be expanded to include the term "Blood" (a child term of  "Body fluid") so files containing either "Body fluid" or  "Blood"  in their metadata will be returned in the search results.  The parent-child relationship is defined by the key "broaders" in the dictionary.  If the full query term is not present in a dictionary then this parameter has no effect.}

\item{use_versions}{Specify which versions of omics data files are used in the query. By default the active version is used. See Versioning above. Syntax:  \\* or \code{v<version number>} or \code{<CHAIN_ID>}:\code{v<version number>} or \code{<CHAIN_ID>}:\code{<accession1,accession2,..>}}

\item{returned_metadata_fields}{The parameter defines amount of metadata attributes to return:  1. \code{template} - return metadata attributes according to applied template, if object doesn’t have applied template default template will be used. This is the default for User endpoints. 2. \code{all} - return all metadata attributes with values and null attributes, if they are present in the applied template. This is the default for Curator endpoints.}

\item{page_offset}{Show the page {pageOffset+1} results from the start of the results. E.g. 100 will show a page of results  starting from the 101st result. The default value is 0.}

\item{page_limit}{Maximum number of results to return per page (see Paging above). This value must be between 0 and 2000 (inclusive). The default is 2000.}
}
\description{
Retrieve groups that match a query
}
